pyproject.toml
[project]
name = "crowd-checker"
version = "0.1.0"
description = "App to check crowd status of particular establishments - For Ateneo MSDS (Ramoncito Jr Bangit)"
authors = [
    {name = "Ramoncito Jr Bangit"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
